add throttling to vitals calls
Before, I had everything in one big Promise.all, but this meant
it was up to JS to decide how many simultaneous requests to
vitals it could make at once. I'm afraid that it'll have adverse
effects on the server trying to do too much work at once, so this
adds throttling such that we only process levels two at a time.

It will of course make syncing take longer, but that doesn't
really matter