add recycle_bin parameter to SQL
when a level is deleted, we're actually going to set the recycle_bin
parameter. the default view exposed does not include levels that have
recycle_bin set to true.

this is mostly motivated by paranoia about a driver possibly breaking
and deleting all the levels by accident.

also there's some wip stuff included in the commit