add/bin/unbin
so if a level isn't present, we don't actually get rid of it completely,
we just set a "recycle_bin" flag. then, if that level comes back, we don't
parse it again, we just unbin it

anyway, rationale here is to act as a 'backup' if any of the sources really
breaks (e.g. someone's google drive gets taken down, etc.)

won't the DB fill up eventually with recycle bin levels? We'll manually
purge it every now and then