add "human name" parameter to sources.txt
So the human name is the name of your "group". Unlike chorus, the
intention is for sources to generally be accessed by multiple
people (e.g. shared google drive folder, discord server, etc.)

and not just an individual folder. in addition, this will hopefully
emphasise that "Rhythm Doctor Lounge" is just another source and
not an official source of levels in any capacity