restructure sql schema
last night I had a dream where I was talking to a giant disembodied
trumpet. the trumpet said that it had patched vitals so that it gave
the running time of levels.

easy, i thought. I'll just reindex the levels. But oh no! The reindexing
process clears the level table, and because the aux table is set to
"ON DELETE CASCADE", that wiped all the verification data!

then everyone in pathology-lab pinged me a hundred times

so the point of this restructuring is to make it so that everything that
needs to be kept in a reindexing is in the aux table and everything that
is derived from the rdzip/sources.txt just stays in the level table.

this way, the process for updating is:

 - dump the aux table
 - reindex
 - reinsert the aux table

and we keep the verification data