starting work on transitioning away from postgrest
why move away from postgrest? it seemed like a good idea
at first, but as I started to use it more I got a better
idea of its limitations. it's really fast to do things
in postgrest that it's good at, but if you have to do
something that postgrest _isn't_ good at, then you're
forced into writing procedural code in your SQL, which is... eww

there are also some... social reasons for it. I don't like how
complex the APIs that postgREST generates are, and a lot
of the routes it generates are completely pointless for orchard
(for instance, postgREST generates endpoints for the aux/status
table, but that's never useful by itself.) and I would like people
to build tools on top of orchard with a straightforward API rather
than the weird postgrest one

in addition, a straight express app is closer to what chorus does,
and it would be nice to be able to contribute back to chorus at some
point in the future