so I'm an idiot
in the Switch component, I used useMemo() to only recompute what element got rendered
if the arguments to the Switch component changed. Makes sense, right?

But what if the props to the _currently rendered component_ change? Then useMemo will
happily keep the version of that component with the _old props_ still around instead
of rerendering the child component! I just spent an hour trying to work this out ahhhh