uh so there was a memory leak in the scraper
and I don't know why it happens! isn't javascript supposed to have a gc that
solves everything for you??

anyway, so the server stopped responding and i couldn't ssh into it at all! so
i had to restart the server each time this occured. but what I discovered was that
this happened roughly every 15 days. huh???

the scraper wasn't releasing the memory from the previous run or something like that,
which caused it to balloon in memory usage. it looks like each run takes about 30 MB of
RAM, which over 15 days caused it to use up all the 32 GB of RAM in the server. at this
point the server crashes, and can't even accept ssh connections because ssh connections
use ram that the server didn't have because the scraper was eating up all the ram ahhhhh

anyway, so what i'm doing here is instead of running the scraper in a while(true) loop
in the code, i'm having it exit after 1 iteration and having Docker restart the container
instead. i'm going to leave this running for a few days and check back to see if the memory
has increased or not